+++ src/lsp/cobol_data/data_types.ml
@@ -21,6 +21,56 @@ open Cobol_common.Srcloc.TYPES
 module NEL = Cobol_common.Basics.NEL
 type 'a nel = 'a NEL.t
 
+(** {2 Literals} *)
+
+type integer_literal =
+  {
+    int_ptree: Cobol_ptree.integer;
+    int_value: integer_value;
+  }
+
+and integer_value = Z.t [@printer Z.pp_print]
+
+and fixed_literal =
+  {
+    fixed_ptree: Cobol_ptree.fixed;
+    fixed_value: fixed_value;
+  }
+
+and fixed_value = Q.t [@printer Q.pp_print]
+
+and floating_literal =
+  {
+    float_ptree: Cobol_ptree.floating;
+    float_value: floating_value;
+  }
+
+and floating_value =
+  {
+    float_significand: fixed_value;
+    float_exponent: int;                    (* 0 <= . <= 9999 in ISO/IEC 2014 *)
+  }
+
+and alphanum_literal = Cobol_ptree.alphanum
+
+and alphanum_value = Cobol_ptree.alphanum
+
+and boolean_literal =
+  {
+    bool_ptree: Cobol_ptree.boolean;
+    bool_value: boolean_value;
+  }
+
+and boolean_value =
+  {
+    bool_width: int;                                (** may be 0 *)
+    bool_bits: Z.t; [@printer Z.pp_print]           (** irrelevant if 0-width *)
+  }
+
+[@@deriving show]
+
+(** {2 Storage} *)
+
 type picture_config = Data_picture.TYPES.config
 type picture = Data_picture.t
 
@@ -72,6 +122,8 @@ type length_variability =
   | Variable_length
   (* Note: OCCURS DYNAMIC is considered fixed-length in ISO/IEC *)
 
+(** {2 Records} *)
+
 type record =
   {
     record_name: string;
@@ -95,9 +147,9 @@ and field_definition =
     field_layout: field_layout;
     field_length_variability: length_variability;
     field_conditions: condition_names; (** Named conditions on the value of this field. *)
-    field_redefines: Cobol_ptree.qualname with_loc option; (** 
-      Set iff this field is a redefinition. 
-      In that case this field appears inside item_redefinitions of the item it redefines. 
+    field_redefines: Cobol_ptree.qualname with_loc option; (**
+      Set iff this field is a redefinition.
+      In that case this field appears inside item_redefinitions of the item it redefines.
       Later, we may create instead a item_redefinition type. *)
     field_redefinitions: item_redefinitions; (** List of alternative definitions for this field *)
     field_has_definition_issues: bool;
@@ -122,8 +174,8 @@ and table_definition =
     table_range: table_range;
     table_init_values: Cobol_ptree.literal with_loc list;     (* list for now *)
     table_redefines: Cobol_ptree.qualname with_loc option; (* same as [field_redefines] but for tables *)
-    table_redefinitions: item_redefinitions; (** 
-      List of alternative definitions for the full table. 
+    table_redefinitions: item_redefinitions; (**
+      List of alternative definitions for the full table.
       Note that by default the typechecker generates a warning on table redefinition. *)
     table_has_definition_issues: bool;
   }