+++ test/lsp/lsp_hover.ml
@@ -14,7 +14,7 @@
 open Lsp.Types
 open Lsp_testing
 
-let print_hovered ?(always_show_hover_definition_text_in_data_div=true)
+let print_hovered ?(show_hover_text_on_definitions = true)
     server ~projdir (prog, prog_positions) =
   let server, prog = add_cobol_doc server ~projdir "prog.cob" prog in
   let location_as_srcloc = new srcloc_resuscitator_cache in
@@ -25,7 +25,7 @@ let print_hovered ?(always_show_hover_definition_text_in_data_div=true)
       position.line position.character;
     match
       LSP.Request.INTERNAL.hover server params
-        ~always_show_hover_definition_text_in_data_div
+        ~show_hover_text_on_definitions
     with
     | None ->
         Pretty.out "Hovering nothing worthy@."
@@ -1484,6 +1484,266 @@ let%expect_test "hover-datadef-table-and-index" =
     ---
     References: 2 |}];;
 
+let%expect_test "hover-preproc-directives" =
+  Unix.putenv "ABCD" "ABCD-VALUE"; (* Warning: left in environment after the test *)
+  let { projdir; end_with_postproc }, server = make_lsp_project () in
+  let prog_n_markers =
+    extract_position_markers {cobol|
+       IDENTIFICATION DIVISION.
+       PROGRAM-ID. prog.
+       DATA DIVISION.
+       WORKING-STORAGE SECTION.
+       >>DEFINE _|_X AS 1
+       >>DEFINE _|_B AS b'10'
+       >>IF B_|_
+       77 WS VALUE "OK".
+       >>ELSE
+       77 WS VALUE "KO".
+       >>END-IF
+       >>DEFINE _|_ABCD AS PARAMETER
+       >>IF _|_ABCD IS DEFINED
+       >>END-IF
+      * Some documentation for bis
+      * ...
+      * on several lines.
+       >>DEFINE _|_BIS AS 42.24
+    |cobol}
+  in
+  print_hovered server ~projdir ~show_hover_text_on_definitions:false
+    prog_n_markers;
+  Pretty.out "Now with hover text on defintions@\n";
+  print_hovered server ~projdir ~show_hover_text_on_definitions:true
+    prog_n_markers;
+  end_with_postproc [%expect.output];
+  [%expect {|
+    {"params":{"diagnostics":[],"uri":"file://__rootdir__/prog.cob"},"method":"textDocument/publishDiagnostics","jsonrpc":"2.0"}
+    (line 5, character 16):
+    __rootdir__/prog.cob:6.16-6.17:
+       3          PROGRAM-ID. prog.
+       4          DATA DIVISION.
+       5          WORKING-STORAGE SECTION.
+       6 >        >>DEFINE X AS 1
+    ----                   ^
+       7          >>DEFINE B AS b'10'
+       8          >>IF B
+    References: 1
+    (line 6, character 16):
+    __rootdir__/prog.cob:7.16-7.17:
+       4          DATA DIVISION.
+       5          WORKING-STORAGE SECTION.
+       6          >>DEFINE X AS 1
+       7 >        >>DEFINE B AS b'10'
+    ----                   ^
+       8          >>IF B
+       9          77 WS VALUE "OK".
+    References: 2
+    (line 7, character 13):
+    __rootdir__/prog.cob:8.12-8.13:
+       5          WORKING-STORAGE SECTION.
+       6          >>DEFINE X AS 1
+       7          >>DEFINE B AS b'10'
+       8 >        >>IF B
+    ----               ^
+       9          77 WS VALUE "OK".
+      10          >>ELSE
+    Compilation variable with value 2
+    ---
+    References: 2
+    (line 12, character 16):
+    __rootdir__/prog.cob:13.16-13.20:
+      10          >>ELSE
+      11          77 WS VALUE "KO".
+      12          >>END-IF
+      13 >        >>DEFINE ABCD AS PARAMETER
+    ----                   ^^^^
+      14          >>IF ABCD IS DEFINED
+      15          >>END-IF
+    References: 2
+    (line 13, character 12):
+    __rootdir__/prog.cob:14.12-14.16:
+      11          77 WS VALUE "KO".
+      12          >>END-IF
+      13          >>DEFINE ABCD AS PARAMETER
+      14 >        >>IF ABCD IS DEFINED
+    ----               ^^^^
+      15          >>END-IF
+      16         * Some documentation for bis
+    Compilation variable with value "ABCD-VALUE" (defined in process environment)
+    ---
+    References: 2
+    (line 18, character 16):
+    __rootdir__/prog.cob:19.16-19.19:
+      16         * Some documentation for bis
+      17         * ...
+      18         * on several lines.
+      19 >        >>DEFINE BIS AS 42.24
+    ----                   ^^^
+      20
+    References: 1
+    {"params":{"diagnostics":[],"uri":"file://__rootdir__/prog.cob"},"method":"textDocument/publishDiagnostics","jsonrpc":"2.0"}
+    Now with hover text on defintions
+    (line 5, character 16):
+    __rootdir__/prog.cob:6.16-6.17:
+       3          PROGRAM-ID. prog.
+       4          DATA DIVISION.
+       5          WORKING-STORAGE SECTION.
+       6 >        >>DEFINE X AS 1
+    ----                   ^
+       7          >>DEFINE B AS b'10'
+       8          >>IF B
+    Compilation variable with value 1.0
+    ---
+    References: 1
+    (line 6, character 16):
+    __rootdir__/prog.cob:7.16-7.17:
+       4          DATA DIVISION.
+       5          WORKING-STORAGE SECTION.
+       6          >>DEFINE X AS 1
+       7 >        >>DEFINE B AS b'10'
+    ----                   ^
+       8          >>IF B
+       9          77 WS VALUE "OK".
+    Compilation variable with value 2
+    ---
+    References: 2
+    (line 7, character 13):
+    __rootdir__/prog.cob:8.12-8.13:
+       5          WORKING-STORAGE SECTION.
+       6          >>DEFINE X AS 1
+       7          >>DEFINE B AS b'10'
+       8 >        >>IF B
+    ----               ^
+       9          77 WS VALUE "OK".
+      10          >>ELSE
+    Compilation variable with value 2
+    ---
+    References: 2
+    (line 12, character 16):
+    __rootdir__/prog.cob:13.16-13.20:
+      10          >>ELSE
+      11          77 WS VALUE "KO".
+      12          >>END-IF
+      13 >        >>DEFINE ABCD AS PARAMETER
+    ----                   ^^^^
+      14          >>IF ABCD IS DEFINED
+      15          >>END-IF
+    Compilation variable with value "ABCD-VALUE" (defined in process environment)
+    ---
+    References: 2
+    (line 13, character 12):
+    __rootdir__/prog.cob:14.12-14.16:
+      11          77 WS VALUE "KO".
+      12          >>END-IF
+      13          >>DEFINE ABCD AS PARAMETER
+      14 >        >>IF ABCD IS DEFINED
+    ----               ^^^^
+      15          >>END-IF
+      16         * Some documentation for bis
+    Compilation variable with value "ABCD-VALUE" (defined in process environment)
+    ---
+    References: 2
+    (line 18, character 16):
+    __rootdir__/prog.cob:19.16-19.19:
+      16         * Some documentation for bis
+      17         * ...
+      18         * on several lines.
+      19 >        >>DEFINE BIS AS 42.24
+    ----                   ^^^
+      20
+    Compilation variable with value 42.24
+    ---
+     Some documentation for bis
+     ...
+     on several lines.
+    ---
+    References: 1
+  |}];;
+
+let%expect_test "hover-datadef-78" =
+  let { projdir; end_with_postproc }, server = make_lsp_project () in
+  print_hovered server ~projdir @@ extract_position_markers {cobol|
+       IDENTIFICATION DIVISION.
+       PROGRAM-ID. prog.
+       DATA DIVISION.
+       WORKING-STORAGE SECTION.
+       78 CO_|_NST VALUE "ABCD".
+      *Note: currently handled as hovering over `01 VAR ... CONST.`
+       77 VAR VALUE CON_|_ST.
+       PROCEDURE DIVISION.
+           DISPLAY "VAR: " _|_VAR ", CONST: " CON_|_ST
+           STOP RUN.
+    |cobol};
+  end_with_postproc [%expect.output];
+  [%expect {|
+    {"params":{"diagnostics":[],"uri":"file://__rootdir__/prog.cob"},"method":"textDocument/publishDiagnostics","jsonrpc":"2.0"}
+    (line 5, character 12):
+    __rootdir__/prog.cob:6.7-6.28:
+       3          PROGRAM-ID. prog.
+       4          DATA DIVISION.
+       5          WORKING-STORAGE SECTION.
+       6 >        78 CONST VALUE "ABCD".
+    ----          ^^^^^^^^^^^^^^^^^^^^^
+       7         *Note: currently handled as hovering over `01 VAR ... CONST.`
+       8          77 VAR VALUE CONST.
+    Compilation variable with value "ABCD"
+    ---
+    References: 3
+    (line 7, character 23):
+    __rootdir__/prog.cob:8.7-8.26:
+       5          WORKING-STORAGE SECTION.
+       6          78 CONST VALUE "ABCD".
+       7         *Note: currently handled as hovering over `01 VAR ... CONST.`
+       8 >        77 VAR VALUE CONST.
+    ----          ^^^^^^^^^^^^^^^^^^^
+       9          PROCEDURE DIVISION.
+      10              DISPLAY "VAR: " VAR ", CONST: " CONST
+    ```cobol
+    VAR
+    ```
+    ```cobol
+    PIC X(4) USAGE DISPLAY
+    ```
+    ALPHANUMERIC(4)
+    VALUE "ABCD"
+    ---
+    Note: currently handled as hovering over `01 VAR ... CONST.`
+    ---
+    References: 2
+    (line 9, character 27):
+    __rootdir__/prog.cob:10.27-10.30:
+       7         *Note: currently handled as hovering over `01 VAR ... CONST.`
+       8          77 VAR VALUE CONST.
+       9          PROCEDURE DIVISION.
+      10 >            DISPLAY "VAR: " VAR ", CONST: " CONST
+    ----                              ^^^
+      11              STOP RUN.
+      12
+    ```cobol
+    VAR
+    ```
+    ```cobol
+    PIC X(4) USAGE DISPLAY
+    ```
+    ALPHANUMERIC(4)
+    VALUE "ABCD"
+    ---
+    Note: currently handled as hovering over `01 VAR ... CONST.`
+    ---
+    References: 2
+    (line 9, character 46):
+    __rootdir__/prog.cob:10.43-10.48:
+       7         *Note: currently handled as hovering over `01 VAR ... CONST.`
+       8          77 VAR VALUE CONST.
+       9          PROCEDURE DIVISION.
+      10 >            DISPLAY "VAR: " VAR ", CONST: " CONST
+    ----                                              ^^^^^
+      11              STOP RUN.
+      12
+    Compilation variable with value "ABCD"
+    ---
+    References: 3
+  |}];;
+
 let%expect_test "hover-datadef-communication-section" =
   let { projdir; end_with_postproc }, server = make_lsp_project () in
   print_hovered server ~projdir @@ extract_position_markers {cobol|
@@ -1522,40 +1782,46 @@ let%expect_test "hover-comment" =
          02 VAL-1 PIC X. *> val1 only inline comment
       * val2 only line comment
          02 VAL-2 PIC X.
+      * val3 several line
+      * comments
+         02 VAL-3 PIC X.
+      * val4 several line
+      * comments and...
+         02 VAL-4 PIC X. *> an inline comment.
        PROCEDURE DIVISION.
-         DISPLAY S_|_TRUCT V_|_AL-1 V_|_AL-2.
+         DISPLAY S_|_TRUCT V_|_AL-1 V_|_AL-2 V_|_AL-3 V_|_AL-4.
          STOP RUN.
     |cobol};
   end_with_postproc [%expect.output];
   [%expect {|
     {"params":{"diagnostics":[],"uri":"file://__rootdir__/prog.cob"},"method":"textDocument/publishDiagnostics","jsonrpc":"2.0"}
-    (line 12, character 18):
-    __rootdir__/prog.cob:13.17-13.23:
-      10         * val2 only line comment
-      11            02 VAL-2 PIC X.
-      12          PROCEDURE DIVISION.
-      13 >          DISPLAY STRUCT VAL-1 VAL-2.
+    (line 18, character 18):
+    __rootdir__/prog.cob:19.17-19.23:
+      16         * comments and...
+      17            02 VAL-4 PIC X. *> an inline comment.
+      18          PROCEDURE DIVISION.
+      19 >          DISPLAY STRUCT VAL-1 VAL-2 VAL-3 VAL-4.
     ----                    ^^^^^^
-      14            STOP RUN.
-      15
+      20            STOP RUN.
+      21
     ```cobol
     STRUCT
     ```
-    Group of 2 subfields
-    Size: 2 bytes
+    Group of 4 subfields
+    Size: 4 bytes
     ---
      inline comment
     ---
     References: 2
-    (line 12, character 25):
-    __rootdir__/prog.cob:13.24-13.29:
-      10         * val2 only line comment
-      11            02 VAL-2 PIC X.
-      12          PROCEDURE DIVISION.
-      13 >          DISPLAY STRUCT VAL-1 VAL-2.
+    (line 18, character 25):
+    __rootdir__/prog.cob:19.24-19.29:
+      16         * comments and...
+      17            02 VAL-4 PIC X. *> an inline comment.
+      18          PROCEDURE DIVISION.
+      19 >          DISPLAY STRUCT VAL-1 VAL-2 VAL-3 VAL-4.
     ----                           ^^^^^
-      14            STOP RUN.
-      15
+      20            STOP RUN.
+      21
     ```cobol
     VAL-1 IN STRUCT
     ```
@@ -1567,15 +1833,15 @@ let%expect_test "hover-comment" =
      val1 only inline comment
     ---
     References: 2
-    (line 12, character 31):
-    __rootdir__/prog.cob:13.30-13.35:
-      10         * val2 only line comment
-      11            02 VAL-2 PIC X.
-      12          PROCEDURE DIVISION.
-      13 >          DISPLAY STRUCT VAL-1 VAL-2.
+    (line 18, character 31):
+    __rootdir__/prog.cob:19.30-19.35:
+      16         * comments and...
+      17            02 VAL-4 PIC X. *> an inline comment.
+      18          PROCEDURE DIVISION.
+      19 >          DISPLAY STRUCT VAL-1 VAL-2 VAL-3 VAL-4.
     ----                                 ^^^^^
-      14            STOP RUN.
-      15
+      20            STOP RUN.
+      21
     ```cobol
     VAL-2 IN STRUCT
     ```
@@ -1585,6 +1851,47 @@ let%expect_test "hover-comment" =
     ALPHANUMERIC(1)
     ---
      val2 only line comment
+    ---
+    References: 2
+    (line 18, character 37):
+    __rootdir__/prog.cob:19.36-19.41:
+      16         * comments and...
+      17            02 VAL-4 PIC X. *> an inline comment.
+      18          PROCEDURE DIVISION.
+      19 >          DISPLAY STRUCT VAL-1 VAL-2 VAL-3 VAL-4.
+    ----                                       ^^^^^
+      20            STOP RUN.
+      21
+    ```cobol
+    VAL-3 IN STRUCT
+    ```
+    ```cobol
+    PIC X USAGE DISPLAY
+    ```
+    ALPHANUMERIC(1)
+    ---
+     val3 several line
+     comments
+    ---
+    References: 2
+    (line 18, character 43):
+    __rootdir__/prog.cob:19.42-19.47:
+      16         * comments and...
+      17            02 VAL-4 PIC X. *> an inline comment.
+      18          PROCEDURE DIVISION.
+      19 >          DISPLAY STRUCT VAL-1 VAL-2 VAL-3 VAL-4.
+    ----                                             ^^^^^
+      20            STOP RUN.
+      21
+    ```cobol
+    VAL-4 IN STRUCT
+    ```
+    ```cobol
+    PIC X USAGE DISPLAY
+    ```
+    ALPHANUMERIC(1)
+    ---
+     an inline comment.
     ---
     References: 2 |}];;
 
@@ -1633,7 +1940,7 @@ let%expect_test "hover-comment-copy" =
 
 let%expect_test "hover-data-division-ref-count-only" =
   let { projdir; end_with_postproc }, server = make_lsp_project () in
-  print_hovered ~always_show_hover_definition_text_in_data_div:false
+  print_hovered ~show_hover_text_on_definitions:false
     server ~projdir @@ extract_position_markers {cobol|
        IDENTIFICATION DIVISION.
        PROGRAM-ID. prog.